mgr/dashboard: move all controller infrastructure to controllers/__init__.py
