+++ src/pybind/mgr/dashboard/controllers/host.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
+from . import ApiController, AuthRequired, RESTController
 from .. import mgr
-from ..tools import ApiController, AuthRequired, RESTController
 
 
 @ApiController('host')
