+++ src/pybind/mgr/dashboard/controllers/monitor.py
@@ -5,8 +5,8 @@ import json
 
 import cherrypy
 
+from . import ApiController, AuthRequired, BaseController
 from .. import mgr
-from ..tools import ApiController, AuthRequired, BaseController
 
 
 @ApiController('monitor')
