+++ src/pybind/mgr/dashboard/controllers/rgw.py
@@ -3,9 +3,9 @@ from __future__ import absolute_import
 
 import json
 
+from . import ApiController, RESTController, AuthRequired
 from .. import logger
 from ..services.ceph_service import CephService
-from ..tools import ApiController, RESTController, AuthRequired
 
 
 @ApiController('rgw')
