+++ src/pybind/mgr/dashboard/controllers/summary.py
@@ -5,10 +5,11 @@ import json
 
 import cherrypy
 
+from . import AuthRequired, ApiController, BaseController
 from .. import logger, mgr
 from ..controllers.rbd_mirroring import get_daemons_and_pools
-from ..tools import AuthRequired, ApiController, BaseController, TaskManager
 from ..services.ceph_service import CephService
+from ..tools import TaskManager
 
 
 @ApiController('summary')
