+++ src/pybind/mgr/dashboard/controllers/tcmu_iscsi.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
+from . import ApiController, AuthRequired, RESTController
 from .. import mgr
 from ..services.ceph_service import CephService
-from ..tools import ApiController, AuthRequired, RESTController
 
 SERVICE_TYPE = 'tcmu-runner'
